backend/requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
python-dotenv==1.1.1
pydantic==2.11.9
Jinja2==3.1.6
# Add any cloud/NLU/payment libs you need (google-adk, stripe, etc.)
